+++ lib/mock-component/mock-component.ts
@@ -1,5 +1,6 @@
 import { core } from '@angular/compiler';
 import {
+  AfterContentInit,
   ChangeDetectorRef,
   Component,
   forwardRef,
@@ -9,12 +10,12 @@ import {
   ViewChild,
   ViewContainerRef,
 } from '@angular/core';
+import { getTestBed } from '@angular/core/testing';
 import { NG_VALUE_ACCESSOR } from '@angular/forms';
 
-import {  MockControlValueAccessor, MockOf } from '../common';
+import { getMockedNgDefOf, MockControlValueAccessor, MockOf } from '../common';
+import { ngMocksUniverse } from '../common/ng-mocks-universe';
 import { directiveResolver } from '../common/reflect';
-
-const cache = new Map<Type<Component>, Type<MockedComponent<Component>>>();
 
 export type MockedComponent<T> = T & MockControlValueAccessor & {
   /** Helper function to hide rendered @ContentChild() template. */
@@ -32,9 +33,13 @@ export function MockComponent<TComponent>(
   component: Type<TComponent>,
   metaData?: core.Directive,
 ): Type<MockedComponent<TComponent>> {
-  const cacheHit = cache.get(component);
-  if (cacheHit) {
-    return cacheHit as Type<MockedComponent<TComponent>>;
+  // we are inside of an 'it'.
+  // It's fine to to return a mock or to throw an exception if it wasn't mocked in TestBed.
+  if ((getTestBed() as any)._instantiated) {
+    return getMockedNgDefOf(component, 'c');
+  }
+  if (ngMocksUniverse.flags.has('cacheComponent') && ngMocksUniverse.cache.has(component)) {
+    return ngMocksUniverse.cache.get(component);
   }
 
   const { exportAs, inputs, outputs, queries, selector } = metaData || directiveResolver.resolve(component);
@@ -89,8 +94,10 @@ export function MockComponent<TComponent>(
     template,
   };
 
+  const config = ngMocksUniverse.config.get(component);
+
   @MockOf(component, outputs)
-  class ComponentMock extends MockControlValueAccessor {
+  class ComponentMock extends MockControlValueAccessor implements AfterContentInit {
     constructor(changeDetector: ChangeDetectorRef) {
       super();
 
@@ -120,10 +127,25 @@ export function MockComponent<TComponent>(
         }
       };
     }
+
+    ngAfterContentInit(): void {
+      if (!(this as any).__rendered && config && config.render) {
+        for (const block of Object.keys(config.render)) {
+          const { $implicit, variables } = config.render[block] !== true ? config.render[block] : {
+            $implicit: undefined,
+            variables: {},
+          };
+          (this as any).__render(block, $implicit, variables);
+        }
+        (this as any).__rendered = true;
+      }
+    }
   }
 
   const mockedComponent: Type<MockedComponent<TComponent>> = Component(options)(ComponentMock as any);
-  cache.set(component, mockedComponent);
+  if (ngMocksUniverse.flags.has('cacheComponent')) {
+    ngMocksUniverse.cache.set(component, mockedComponent);
+  }
 
   return mockedComponent;
 }